pyproject.toml
[project]
name = "mcpclient"
version = "2.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "rich",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "pytest-cov>=7.0.0",
]
